Move functions to the module where they're used
